Fixed a small fix
The fix(previous commit) was to seperate the users using a '|'(or anything that isn't blank space) as people with a name such as "Testing this" equaled to 2(in length) because of the empty space I use to seperate. The game should now ONLY play once 2 people do the same command. Also added in for first person, a message of how to join in.